Dengue fever individual-level reporting data from Puerto Rico (USA) 1990-2010

Surveillance data from CDC Division of Vector-Borne Diseases. 1990-2010 case reporting data included.

Usage

data(denguedat)

Format

A data frame.

Source

Data originally from the NobBS R package. https://cran.r-project.org/package=NobBS

Details

Each row represents a case with the columns indicating the following:

onset_week: the first day of the week of symptom onset, formatted in ISO8601 as YYYY-MM-DD.

report_week: the first day of the week of case report, formatted in ISO8601 as YYYY-MM-DD.

Note

Data originally from the NobBS package. A synthetic gender column was removed to retain only the real surveillance data.

References

MCGOUGH, Sarah F., et al. Nowcasting by Bayesian Smoothing: A flexible, generalizable model for real-time epidemic tracking. PLoS computational biology, 2020, vol. 16, no 4, p. e1007735.

See also

Examples

data(denguedat)