Ebola virus disease outbreak in Sierra Leone, 2014–2015

Line-list of confirmed and suspected Ebola virus disease cases from Sierra Leone from 2014-05-18 to 2015-09-13. Data from the outbreaks package.

Usage

data(ebola_sierraleone_2014)

Format

A data frame with columns:

id

Numeric case identifier.

age

Age of the case (years).

sex

Sex of the case ("F" or "M").

status

Case classification: "confirmed" or "suspected".

date_of_onset

Date of symptom onset

date_of_sample

Date the diagnostic sample was collected.

district

District of the case.

chiefdom

Chiefdom.

A data frame.

Source

Data from the outbreaks R package. https://github.com/reconhub/outbreaks

References

L. Fang, et al. 2016. Ebola virus disease in Sierra Leone. Proceedings of the National Academy of Sciences, 113 (16) 4488-4493; DOI: 10.1073/pnas.1518587113

See also

Author

Data from Fang et al. (2016) Transfer to R and documentation by Finlay Campbell (finlaycampbell93@gmail.com).

Examples

data(ebola_sierraleone_2014)